:recycle:refactor: refactor book detail page 타입 위치 변경 사용하지 않는 파일 삭제 Related to: #269
