+++ src/features/BookDetailPage/types/types.ts
@@ -43,8 +43,42 @@ export interface DbUserReview {
   isbn: string;
 }
 
-export interface BookPostCount {
+export interface DbBookPostCount {
   isbn: string;
   review_count: number;
   posting_count: number;
+}
+
+export interface DbReaderStats {
+  isbn: string;
+  total_collectors: number;
+  demographics: {
+    gender: {
+      male: {
+        '10s': number;
+        '20s': number;
+        '30s': number;
+        '40s': number;
+        '50s': number;
+        '60plus': number;
+        unknown: number;
+      };
+      female: {
+        '10s': number;
+        '20s': number;
+        '30s': number;
+        '40s': number;
+        '50s': number;
+        '60plus': number;
+        unknown: number;
+      };
+      unknown: number;
+    };
+  };
+}
+
+export interface DbBookReviewStats {
+  isbn: string;
+  reviewer_count: number;
+  average_rating: number;
 }
